+++ balances/get_liquidity.js
@@ -1,15 +1,21 @@
 const asyncAuto = require('async/auto');
 const {authenticatedLndGrpc} = require('ln-service');
 const {getChannels} = require('ln-service');
-const {percentile} = require('stats-lite');
+const {getNode} = require('ln-service');
+const {getWalletInfo} = require('ln-service');
 const {returnResult} = require('asyncjs-util');
 
 const balanceFromTokens = require('./balance_from_tokens');
+const {getNetwork} = require('./../network');
+const {getScoredNodes} = require('./../network');
+const liquidityTokens = require('./liquidity_tokens');
 
 const {round} = Math;
 const topPercentile = 0.9;
 
 /** Get the channel available liquidity
+
+  A request function is requried
 
   {
     [above]: <Tokens Above Tokens Number>
@@ -17,6 +23,9 @@ const topPercentile = 0.9;
     [is_outbound]: <Return Outbound Liquidity Bool>
     [is_top]: <Return Top Liquidity Bool>
     lnd: <Authenticated LND gRPC API Object>
+    [min_node_score]: <Minimum Node Score Number>
+    [max_fee_rate]: <Max Inbound Fee Rate Parts Per Million Number>
+    [request]: <Request Function>
     [with]: <Liquidity With Specific Node Public Key Hex String>
   }
 
@@ -30,8 +39,16 @@ module.exports = (args, cbk) => {
     return asyncAuto({
       // Check arguments
       validate: cbk => {
+        if (!!args.is_outbound && args.max_fee_rate !== undefined) {
+          return cbk([400, 'MaxLiquidityFeeRateNotSupportedForOutbound']);
+        }
+
         if (!args.lnd) {
           return cbk([400, 'ExpectedLndToGetLiquidity']);
+        }
+
+        if (!!args.min_node_score && !args.request) {
+          return cbk([400, 'ExpectedRequestFunctionToFilterByNodeScore']);
         }
 
         return cbk();
@@ -42,38 +59,72 @@ module.exports = (args, cbk) => {
         return getChannels({lnd: args.lnd}, cbk);
       }],
 
-      // List of tokens to sum
-      tokens: ['getChannels', ({getChannels}, cbk) => {
-        const activeChannels = getChannels.channels
-          .filter(n => !!n.is_active)
-          .filter(n => !args.with || n.partner_public_key === args.with);
+      // Determine which network the node is on
+      getNetwork: ['validate', ({}, cbk) => getNetwork({lnd: args.lnd}, cbk)],
 
-        const balanceType = !!args.is_outbound ? 'local' : 'remote';
+      // Get the node's public key
+      getNodeKey: ['validate', ({}, cbk) => {
+        return getWalletInfo({lnd: args.lnd}, cbk);
+      }],
 
-        const tokens = activeChannels.map(n => n[`${balanceType}_balance`]);
-
-        if (!!args.is_top) {
-          return cbk(null, [round(percentile(tokens, topPercentile))]);
+      // Get policies
+      getPolicies: ['getNodeKey', ({getNodeKey}, cbk) => {
+        if (args.max_fee_rate === undefined) {
+          return cbk(null, {channels: []});
         }
 
-        return cbk(null, tokens);
+        return getNode({
+          lnd: args.lnd,
+          public_key: getNodeKey.public_key,
+        },
+        cbk);
+      }],
+
+      // Get node scores
+      getScores: ['getNetwork', ({getNetwork}, cbk) => {
+        if (!args.min_node_score) {
+          return cbk(null, {});
+        }
+
+        return getScoredNodes({
+          network: getNetwork.network,
+          request: args.request,
+        },
+        cbk);
+      }],
+
+      // List of tokens to sum
+      tokens: [
+        'getChannels',
+        'getNodeKey',
+        'getPolicies',
+        'getScores',
+        ({getChannels, getNodeKey, getPolicies, getScores}, cbk) =>
+      {
+        return cbk(null, liquidityTokens({
+          channels: getChannels.channels,
+          is_outbound: args.is_outbound,
+          is_top: args.is_top,
+          max_fee_rate: args.max_fee_rate,
+          min_node_score: args.min_node_score,
+          nodes: getScores.nodes,
+          policies: getPolicies.channels.map(n => n.policies),
+          public_key: getNodeKey.public_key,
+          with: args.with,
+        }));
       }],
 
       // Total balances
       total: ['tokens', ({tokens}, cbk) => {
-        const {above} = args;
-        const {below} = args;
-
-        const balance = balanceFromTokens({above, below, tokens});
-
-        return cbk(null, balance);
-      }],
-
-      // Liquidity
-      liquidity: ['total', ({total}, cbk) => {
-        return cbk(null, {balance: total});
+        return cbk(null, {
+          balance: balanceFromTokens({
+            tokens,
+            above: args.above,
+            below: args.below,
+          }),
+        });
       }],
     },
-    returnResult({reject, resolve, of: 'liquidity'}, cbk));
+    returnResult({reject, resolve, of: 'total'}, cbk));
   });
 };