+++ test/fixtures/index.js
@@ -1,6 +1,7 @@
 const chanInfoResponse = require('./chan_info_response');
 const getInfoResponse = require('./get_info_response');
 const getNodeInfoResponse = require('./get_node_info_response');
+const liquidityChannelsResponse = require('./liquidity_channels_response');
 const listChannelsResponse = require('./list_channels_response');
 const listPeersResponse = require('./list_peers_response');
 const pendingChannelsResponse = require('./pending_channels_response');
@@ -9,6 +10,7 @@ module.exports = {
   chanInfoResponse,
   getInfoResponse,
   getNodeInfoResponse,
+  liquidityChannelsResponse,
   listChannelsResponse,
   listPeersResponse,
   pendingChannelsResponse,